backend/requirements.txt
# FastAPI i zaleznosci
fastapi==0.121.3
starlette==0.50.0
uvicorn==0.38.0
h11==0.16.0
click==8.3.1

# Pydantic i zaleznosci
pydantic==2.12.4
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0

# SQLAlchemy i baza danych
SQLAlchemy==2.0.44
asyncpg==0.30.0
greenlet==3.2.4

# Async i networking
anyio==4.11.0
sniffio==1.3.1
idna==3.11

# Konfiguracja
python-dotenv==1.2.1

# Testy
pytest==9.0.1
pytest-asyncio==1.3.0
httpx==0.28.1
httpcore==1.0.9
certifi==2025.11.12
iniconfig==2.3.0
pluggy==1.6.0
packaging==25.0
